Update move card to cards
